+++ tests/unit/neptune/new/internal/utils/test_disk_utilization.py
@@ -18,6 +18,7 @@ import unittest
 import warnings
 from io import UnsupportedOperation
 
+import pytest
 from mock import (
     MagicMock,
     patch,
@@ -31,7 +32,12 @@ from neptune.envs import (
     NEPTUNE_MAX_DISK_UTILIZATION,
     NEPTUNE_NON_RAISING_ON_DISK_ISSUE,
 )
-from neptune.internal.utils.disk_utilization import ensure_disk_not_overutilize
+from neptune.exceptions import NeptuneMaxDiskUtilizationExceeded
+from neptune.internal.utils.disk_utilization import (
+    NonRaisingErrorHandler,
+    RaisingErrorHandler,
+    ensure_disk_not_overutilize,
+)
 
 
 class TestDiskUtilization(unittest.TestCase):
@@ -95,7 +101,7 @@ class TestDiskUtilization(unittest.TestCase):
             disk_usage_mock.side_effect = error
 
             wrapped_func()  # asserting is not required as expecting that any error will be caught
-            mocked_func.assert_not_called()
+            mocked_func.assert_called_once()
 
     @patch.dict(os.environ, {NEPTUNE_NON_RAISING_ON_DISK_ISSUE: "True"})
     @patch.dict(os.environ, {NEPTUNE_MAX_DISK_UTILIZATION: "100"})
@@ -128,7 +134,71 @@ class TestDiskUtilization(unittest.TestCase):
         disk_usage_mock.return_value.percent = 99
         mocked_func = MagicMock()
         wrapped_func = ensure_disk_not_overutilize(mocked_func)
+        with pytest.raises(NeptuneMaxDiskUtilizationExceeded):
+            wrapped_func()
 
-        wrapped_func()
+        mocked_func.assert_not_called()
 
-        mocked_func.assert_called_once()
+
+class TestDiskErrorHandler(unittest.TestCase):
+    @patch("neptune.internal.utils.disk_utilization.RaisingErrorHandler")
+    @patch("neptune.internal.utils.disk_utilization.NonRaisingErrorHandler")
+    @patch.dict(os.environ, {NEPTUNE_NON_RAISING_ON_DISK_ISSUE: "False"})
+    def test_raising_handler_used_if_env_var_false(self, mock_non_raising_handler, mock_raising_handler):
+        decorated = ensure_disk_not_overutilize(MagicMock())
+        decorated()
+        mock_raising_handler.assert_called_once()
+        mock_non_raising_handler.assert_not_called()
+
+    @patch("neptune.internal.utils.disk_utilization.RaisingErrorHandler")
+    @patch("neptune.internal.utils.disk_utilization.NonRaisingErrorHandler")
+    @patch.dict(os.environ, {NEPTUNE_NON_RAISING_ON_DISK_ISSUE: "True"})
+    def test_non_raising_handler_used_if_env_var_true(self, mock_non_raising_handler, mock_raising_handler):
+        decorated = ensure_disk_not_overutilize(MagicMock())
+        decorated()
+        mock_non_raising_handler.assert_called_once()
+        mock_raising_handler.assert_not_called()
+
+    def test_non_raising_handler(self):
+        func = MagicMock()
+        func.side_effect = OSError
+
+        handler = NonRaisingErrorHandler(None, func)
+        handler.handle_limit_not_set()  # should not raise exception
+
+        handler = NonRaisingErrorHandler(90.0, func)
+        handler.handle_utilization_calculation_error()  # should not raise exception
+
+        handler.handle_limit_not_exceeded()  # should not raise exception
+
+        handler.handle_limit_exceeded(100)  # should not raise exception
+
+        handler.run()  # should not raise exception
+
+    def test_raising_handler(self):
+        func = MagicMock()
+        func.side_effect = OSError
+
+        with pytest.raises(OSError):
+            handler = RaisingErrorHandler(None, func)
+            handler.handle_limit_not_set()
+
+        with pytest.raises(OSError):
+            handler = RaisingErrorHandler(None, func)
+            handler.handle_utilization_calculation_error()
+
+        with pytest.raises(OSError):
+            handler = RaisingErrorHandler(100.0, func)
+            handler.handle_limit_not_exceeded()
+
+        with pytest.raises(NeptuneMaxDiskUtilizationExceeded):
+            handler = RaisingErrorHandler(90.0, func)
+            handler.handle_limit_exceeded(100)
+
+        with pytest.raises(OSError):
+            handler.run()
+
+        func.side_effect = None
+        with patch("neptune.internal.utils.disk_utilization.get_disk_utilization_percent", return_value=95):
+            with pytest.raises(NeptuneMaxDiskUtilizationExceeded):
+                handler.run()